# Faster column creation when building subspaces