Original timings (truncated):

> %Own   %Total  OwnTime  TotalTime  Function (filename:line)
  0.00%   0.00%   0.000s    303.8s   <module> (ts_ihacres_adaptive_leja_pce.py:93)                                                                                                                                          
  0.00%   0.00%    8.16s    90.29s   add_columns_to_pivoted_lu_factorization (pyapprox\utilities.py:867)                                                                                                                    
  0.00%   0.00%   59.04s    59.04s   swap_rows (pyapprox\utilities.py:896)                                                                                                                                                  
  0.00%   0.00%    3.54s    49.34s   add_columns_to_pivoted_lu_factorization (pyapprox\utilities.py:870)   

As indicated in the above, the original time taken for my example was ~300 seconds (see top row)
59 seconds of this was spent swapping rows when adding new subspaces (see third row)

After improvements, this was brought down to ~240 seconds, with much time spent in `np.outer`. This is a little misleading, it's actually the loop surrounding the call to `np.outer`

> %Own   %Total  OwnTime  TotalTime  Function (filename:line)
  0.00%  80.00%   0.000s    241.8s   <module> (ts_ihacres_adaptive_leja_pce.py:93)                                                                                                                                                                                                                                                            
  1.00%   1.00%   35.89s    81.20s   add_columns_to_pivoted_lu_factorization (pyapprox\utilities.py:874)                                                                                                                    
  0.00%   0.00%   0.160s    45.85s   outer (<__array_function__ internals>:5)                                                                                                                                               
  0.00%   0.00%   45.38s    45.38s   outer (numpy\core\numeric.py:942)

If we cythonize `swap_rows` and critical paths in `add_columns_to_pivoted_lu_factorization`, this can be brought down to ~180 seconds. The cythonized version is not part of this PR.

>  %Own   %Total  OwnTime  TotalTime  Function (filename:line)
  0.00% 100.00%   0.000s    183.1s   build (pyapprox_interface\pce_interface.py:131)
 39.00%  49.00%   86.42s    126.1s   add_columns_to_pivoted_lu_factorization (pyapprox\utilities.py:850)